Create Which-Game-is-Winning Rate-on-Mahadev-Book.md
It is usual that the users of Mahadev Book search the best game with the most significant winning percentage. As more and more individuals register on the platform on a daily basis, several users want to get the best out of it. But which is the game that will give you the best investment of your time and efforts?